User kernel-side pyflyby to insert and refomat imports.
This fixes issues where import could be added in the middle of multiline
imports. I am not too sure if this will be have correctly in all the
case, in particular if there is multiple imports at once, but that might
need to be fixed in pyflyby itself to send a single mesage in the comms
with the list of imports to add.