Fix imports from long running cells not added
This PR improves the lock mechanism implemented in PR#7.

Instead of starting timers immediately as the lock is acquired, we wait
till the previous lock expires to start the timeout for the next.

Also, if the current timeout expires when the kernel is busy, we restart
it because the kernel is most likely executing code and we cannot
process the message at that point.

Finally, the timeout of the locks is set to 999999 secs by default which
essentially disables the locks by default and most users shouldn't need
to change this setting. This is required to be lowered only when in a
lossy/slow network. When a lock expires, the current import is skipped
and another message is sent to the kernel to resolve the next import.